Start SphericalMercator module
Used to translate from centre of tile (x,y) in metres to a bounding box
of lower-left and upper-right points for some zoom level.

Turns out what we actually want is to go from Google tile coordinates at
a given zoom level, to a bounding box in lat/lon within spherical
mercator. We need lat/lon to put in our PostGIS queries to limit the
features we return to those that intersect a bounding box.