WIP flesh out mvt handler
Start filling in how to get a tile from a request. Can't seem to get the
parameter decoding quite right for Hasql, but the rest seems to be OK.

Now require that user queries **always** return 2 columns: the GeoJSON
and a HStore containing a map of properties to their values. That way we
can (hopefully!) easily decode the HStore to a `Map Text Text`, which we
can turn into JSON as part of our input to Mapnik.